Fix comments after code rview
