+++ Lesson08/OperationEnum.cs
@@ -1,0 +1,8 @@
+﻿namespace Lesson08;
+
+public enum OperationEnum
+{
+    Sum = 1,
+    MaxMin = 2,
+    Average = 3
+}
