+++ Lesson08/PersonConstants.cs
@@ -1,0 +1,6 @@
+﻿namespace Lesson08;
+
+public class PersonConstants
+{
+    public static string AdditionalMessage = nameof(AdditionalMessage);
+}
